Add MBC1 passing firsts MoonEye tests | change all addr type from uint16_t to size_t to avoid overflow with bigger ROM

+++ Emulator/IMemoryRange.hpp
@@ -3,9 +3,9 @@
 class IMemoryRange
 {
 public:
-	virtual uint8_t read8(uint16_t addr) = 0;
-	virtual void write8(uint16_t addr, uint8_t v) = 0;
-	virtual bool isInRange(uint16_t addr) const = 0;
+	virtual uint8_t read8(size_t addr) = 0;
+	virtual void write8(size_t addr, uint8_t v) = 0;
+	virtual bool isInRange(size_t addr) const = 0;
 	virtual size_t size() const = 0;
 	virtual const char* name() const = 0;
 };
